add session verification for new.php and delete.php

+++ delete.php
@@ -3,13 +3,19 @@ include 'connection.php';
 
 session_start();
 
-$query = 'DELETE FROM post_it WHERE id=:id';
-$response = $bdd->prepare($query);
-$response->execute([
-    'id' => $_GET['id']
-]);
+if (isset($_SESSION['user'])) {
+    $query = 'DELETE FROM post_it WHERE id=:id';
+    $response = $bdd->prepare($query);
+    $response->execute([
+        'id' => $_GET['id']
+    ]);
+    
+    header('location: index.php');
+    exit();
+} else {
+    header('location: index.php');
+    exit();
+}
 
-header('location: index.php');
-exit();
 
 ?>
